zsh major updates
Add documentation module.
Document all liable functions.
Add completion to d().
Add daemon listing to d().
Add XDG module.
Change all tests to use internal (change [ to [[)
Remove vim -p alias.
Add module listing support.
Add chpwd module.
Add git graph aliases.
Add git branch setup.
Add git action CVS prompts.
Remove the global dotdot aliases. They were unused anyway.
Make prompt dynamic until manually changed. Remove PKEEP modification.

+++ zsh/modules/apps/django.zsh
@@ -8,8 +8,21 @@ alias dz='bpython'
 # Django sync hard. Useful when you update a model and the regular sync
 # cant catch it. Note that extensive fixtures are crucial for this to be
 # useful.
-function dsh()
+function dsh()##
 {
+	if [[ "$1" = "--zdoc" ]] ; then
+		if [[ "$2" =~ "s(hort)?" ]] ; then
+			echo "Resync a Django database"
+		elif [[ "$2" =~ "l(ong)?" ]] ; then
+			local str="Drops all mysql database tables inside ´dev_main´\n"
+			str+="and reloads all fixtures. Useful when editing existing\n"
+			str+="models and the regular sync won't catch it.\n\n"
+			str+="Note that without fixtures, this just kills all your data."
+			echo $str
+		fi
+		return
+	fi
+
 	for db in `mysql -s -s -e "show tables in dev_main"` ; do
 		mysql -e "drop table dev_main.$db"
 	done
